Tests sin

+++ tinygrad/function.py
@@ -41,14 +41,14 @@ class Sin(Function):
     # normalize x with analogue of math.fmod(x, 2*pi), the double cast is to get the floor
     y = x.cast(dtypes.float64)
     y = y.e(BinaryOps.SUB, y.e(BinaryOps.DIV, y.const(math.pi*2)).cast(dtypes.int).cast(dtypes.float64).e(BinaryOps.MUL, y.const(math.pi*2)))
-    self.x = ret = acc = y
+    self.x = ret = acc = y.cast(x.dtype)
     self.precision = 20
     #if x.dtype == dtypes.float64:
     #   self.precision = 25
     for i in range(1, self.precision):
       acc = acc.e(UnaryOps.NEG).e(BinaryOps.MUL, self.x).e(BinaryOps.MUL, self.x)
       ret = ret.e(BinaryOps.ADD, acc.e(BinaryOps.DIV, self.x.const(math.factorial(2*i + 1))))
-    return ret
+    return ret.cast(x.dtype)
 
   def backward(self, grad_output:LazyBuffer) -> LazyBuffer:
     ret = acc = self.x.const(1)
